Adjust documentation: PACKAGE_USAGE_STATS will only be granted to system apps
As so, UsageStatsManager won't be an alternative for regular apps.

NOTE: This API requires the permission android.permission.PACKAGE_USAGE_STATS, which is a system-level permission and will not be granted to third-party apps. However, declaring the permission implies intention to use the API and the user of the device can grant permission through the Settings application.